+++ src/main/java/calculator/machine/CalculatingMachine.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.machine;
 
 import java.util.List;
