+++ src/main/java/calculator/separator/CustomSeparator.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.separator;
 
 import java.util.ArrayList;
 import java.util.List;
